Use the new partner customization to load wallpapers
Lollipop introduced the ability to load 3rd party wallpapers provided
that those apks are system apps.  All that is required is to declare
a receiver in the manifest that handles the action
com.android.launcher3.action.PARTNER_CUSTOMIZATION.  This receiver
is never launched so there is no need to have a source file for it.

Change-Id: I967b75f30785710e59926783dc018a20b0fd8625

Add back no wallpaper option

Change-Id: I66469489973af60dbb601324e1bcc89d22b063f3